Refactor deposit locktime handling
It makes more sense to have the `refundLocktime`
directly in the `Deposit` interface instead of
computing it during the deposit script construction.
This is because the locktime will be used in
multiple places.
